requirements.txt
beautifulsoup4==4.12.2
fake_useragent==1.4.0
jaro_winkler==2.0.3
pandas==2.1.4
Pillow==10.1.0
Pillow==10.1.0
Requests==2.31.0
setuptools==65.5.1
